resume-analyzer/requirements.txt
streamlit
google-genai
pdfplumber
docx2txt
requests
beautifulsoup4
reportlab